Fixed build errors
The CMake file needed to be updated to include the sqlite package.

And the buffer need to be referenced in the ClipboardHelper.cpp file to tells the compiler that there will be only one instance of buffer shared across all instances of the ClipboardHelper class, and it should be defined and initialized in one of your source files (e.g., ClipboardHelper.cpp).